Ability to delete Integrantes

+++ src/main/java/com/jeep/lolesports/dao/IntegranteDao.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface IntegranteDao {
     List<Integrante> findAll();
-    Integrante findById(Long id);
+    Integrante findById(int id);
     void save(Integrante integrante);
     void delete(Integrante integrante);
 }
